This appears never to have been a good idea
javaNameString was passed in as the type name from the explicit
method signature.

This line could only do one of two things: replace it with the
same name, and be correct, or replace it with anything else,
and be wrong.

Ok, "never" hasn't been very long. I introduced it in ffe41c8.
It was modeled on the C buildSignature code, which used
Type_getJNISignature, and substituted Type_getJavaTypeName as
needed to work with method handles in Java.

Turns out Type_getJNISignature has an overridden implementation
for a Coerce type (returns the signature of the Coerce's inner
type), while Type_getJavaTypeName was never taught the same trick.

Still, I'm not convinced it makes sense to do anything here, at
most assert that the Type's name IS the one from the explicit
signature.