DDRProcessor using Identifier.Qualified for types
Had to introduce DBType, an abstraction over database type names,
which are usually qualified identifiers, but sometimes reserved
SQL syntax. DBType also has to hold stuff like the [] suffix on
an array type, which clearly isn't part of the qualified name.
It is also currently being used to hold the defaults string when
returned from typeWithDefault().

DBType clearly has room to improve. It doesn't have equals() right
now, as it would be hard to do the right thing except in the case of
a plain Identifier.Qualified without array suffix or defaults. That
might be barely ok for the purpose of recognizing user-defined types,
as those will probably always have qualified identifier names, not
SQL reserved syntax, and won't be replacing arrays. Still, it should
probably handle arrayness more explicitly than just as a string
suffix, and maybe should not be overloaded for defaults duty (or just
define equals() to be independent of any defaults payload it might
be carrying).