DBType now groks modifier, array, and default
It now has fromSQLTypeAnnotation to construct a DBType from
whatever could be found in the 'value' element of SQLType, which
formerly could be anything PostgreSQL wouldn't object to when
dumped into a deployment descriptor where a type should go. It
is still about that lenient, but will try to tease out a qualified
identifier for the type whenever it can make enough sense of the
syntax to do so. If it has done so, then dependTag() will return
a DependTag.Type for it.

If the syntax did not fit the generic qualified-identifier,
optional parenthesized typmod, optional array dimensions form, it
will have been treated as a type named with SQL's grammatically
reserved syntax, and dependTag will return null. That should be
adequate, as the only types named using grammatically reserved
SQL syntax should be those that are already known to be present,
and therefore uninteresting for the task of ordering the deployment
descriptor.

It may be overkill to parse the SQLType annotation value with all
the machinery in Lexicals, such as allowing separator() in all the
places it could go in SQL. It would have been possible to choose
syntax restrictions for an SQLType value (no spaces around the
qualified identifier dot! no comments!) and hand-roll shorter
parsing code tailored to it. But it is good to begin getting some
experience applying the Lexicals machinery to real work; the full
capabilities will be needed for future tasks like overhauling the
deployment-time descriptor parsing.