Simplify
As long as there is a fallback policy to parse as Reserved whenever
what follows the initial identifier (and precedes array dims, if any)
does not look like a parenthesized type modifier, then there is no need
for extra work to recognize DOUBLE PRECISION, NATIONAL CHARACTER or
NATIONAL CHAR as reserved; the second word will trigger the fallback
treatment. If the fallback policy is ever changed (say, to report such
cases as clearly erroneous), then the extra DOUBLE/NATIONAL checking
removed in this commit will have to go back in.

In the spirit of simplification, use the existing pattern
ISO_AND_PG_IDENTIFIER_CAPTURING here rather than deriving a new one
that excludes the Unicode-delimited-identifier case. It is true that
the input values here are coming from Java source, which already has
its own convention for escaping Unicode characters, and there could
be an argument for not doing extra work to support an additional
convention for accomplishing the same thing. But that may not be a
strong argument for doing extra work NOT to support it. As prior
PL/Java versions dumped the SQLType string directly into the
descriptor for PostgreSQL to make sense of, allowing all of the
identifier styles here reduces the chance of breaking code that worked.

This change involved revisiting the ISO_UNICODE_ESCAPE_SPECIFIER pattern
to make sure it's right, and expanding a comment to confirm that it is.