Record dependency tags for functions
A Function now records that it provides itself, and requires
its parameter and return types. (Nothing useful happens because
of that, until types also record that they provide themselves).

DependTag.Function had to be extended to identify the function
by parameter-type signature as well as by name. Overlooked that
at first.

That in turn required a very rudimentary equals() implementation
for DBType. It may well be unsuited for any use other than
function signature comparison, where defaults and modifiers don't
matter. It will also be very easy to see false negatives for
DBType.Reserved owing to semantically meaningless differences in
spelling of the uninterpreted reserved text. (This could be a reason
to revisit the way Reserved currently makes no even halfhearted effort
to segregate 'name' from 'modifier'.)

But false negatives of that sort might turn out to be tolerable.
They won't lead to more-than-one-provider errors as false positives
would, and implicit tags generated from the same sources will probably
be spelled the same. If they aren't, they won't match up at
toplogical ordering time, and will just be ignored, with the possible
worst-case result that explicit tags will have to be added to avoid
an ordering mistake. That doesn't sound like the end of the world.