Record provideTags for MappedUDT
And remove some no-longer-needed explicit dependency annotations
from the ComplexTuple example.

In passing, tighten the cycle reports from the topological orderer,
and give AbstractAnnotationImpl a serviceable toString() for
debugging convenience.

BaseUDT is going to be interesting: the inherent cycle between type
and type functions can be escaped in the install direction by splitting
the type creation into the shell and the final update. That trick isn't
available for the remove direction, which will require some combination
of DROP TYPE ... CASCADE and suppression of the undeploy strings that
would otherwise be generated by the type functions.