Suppress qualified-export warning in pljava-api
The annotation.processing package isn't really deemed
polished enough to call API, so it isn't generally exported,
but for now it is exported to o.p.p.internal because
InstallHelper uses a DDRWriter method. That's kind of ugly
and should be fixed eventually by moving a more polished
version of that functionality into the o.p.p.sqlgen package,
which is exported unconditionally, and letting the internal
module use that.

There remain some warnings in DDRProcessor, but that's being
refactored in another branch.