Add a variadic-call test to AnyTest
It makes sense to add it there because of a current bug in Type
caching that creates an order dependency between this test and
AnyTest's makeArray method: a Heisenbuggy behavior depending on
what type oid was last associated in the cache with the Object[]
type.

Other than that bug, it appears not at all difficult to support
variadic functions. PostgreSQL takes care of gathering the
arguments into an array, so that while the Java method may be
declared variadic, it does not have to be; it may simply declare
an array parameter in the right position. This is desirable,
because a declared variadic parameter would need to be the last one
in the Java method's argument list, which conflicts with the existing
convention for composite returns. But it is no problem to simply
declare a variadic parameter in the SQL function declaration, and
a corresponding ordinary array parameter in the Java method.

Completing variadic support will require fixing the Type caching bug
and creating some annotation syntax to tell the DDR generator to
produce the keyword VARIADIC. Future work; this test is simply here
to make sure support does not regress meanwhile.