Simplify the Eclipse-specific stuff in pom.xml
For Maven2Eclipse to know what to do with the various goals
of a plugin, it was once necessary to have a pluginManagement
section like the long one removed here. In M2E 1.1, it became
possible for the plugin authors to include that information in
the jar of a plugin. That is already done by the nar-maven-plugin
and the properties-maven-plugin (at version 1.0.0, to which it is
upgraded here). That leaves only the maven-antrun-plugin needing
special treatment.

M2E 1.7 introduced a much less obtrusive form of special treatment,
needing only the <?m2e ignore?> processing instruction added here.

However (nothing is ever simple), that requires declaring Maven 3.5.2
the minimum needed to build the project. Processing instructions made
earlier versions crash (MNG-6280).

So this can't be backpatched to REL1_5_STABLE; 3.5.2 requires Java 7.