Update doc of oldest usable Maven version
As reported in #207, owing to incompatible later versions of
plugins we haven't pinned, Mavens as old as 3.0.5 or so are
no longer usable.

As of today, 3.2.5 will still complete a build starting with an
empty repository, and 3.2.5 is as old as anyone should need; it
will run on Java 6.

The larger issue that we should be more systematic about pinning
dependency versions has merit, but it's, well, a larger issue.