Get examples javadoc working again
The examples are still non-modularized code (as any PL/Java
user code will be, for 1.6), which relies on the PL/Java API,
which is modular. That's absolutely a supported situation that
was explicitly planned for in the JPMS design to allow for
orderly migration. But javadoc didn't get the memo (and apparently
won't until Java 15):
https://bugs.openjdk.java.net/browse/JDK-8240169

Meanwhile, of various workarounds sampled from the internets,
downgrading the maven-javadoc-plugin version to 3.0.1, only
for the -examples subproject, seems to be the only one that
works here. Credit to this issue in the jOOQ project, where
they seem to have gone through the same stages of grief:
https://github.com/jOOQ/jOOQ/issues/9099