Reactivate links to Java API
All day yesterday as I was occupied with the module-source-path
issue, the javadoc plugin was adding options to link to Java API
without being asked.

Today it is no longer doing that, and you can't pay me enough
to bother finding out why. Instead, just turn off the detection
and specify the option manually. For this, there are config options.

This has the arguable advantage of producing links to a consistent
version of the Java API. I would like that to be the Java 9 API,
but apparently the package-list / element-list business didn't
get sorted on Oracle's end until 10. So link that version; it
seems to be the last one that doesn't require hardcoding a language
tag into the URL.