Update maven-project-info-reports-plugin etc.
This branch is turning into more of a "things that go bump while
building docs" branch than merely a "javadoc plugin" branch.

Bump the maven-project-info-reports-plugin to 3.0.0 because its
"dependencies" analysis uses BCEL, and the older version of that
was flooding out errors on newer class files.

... which breaks version 3.4 of maven-site-plugin, so update that
to 3.9.0.

... which breaks ${project.*} references in site.xml, so edit those.

And then observe sweet success.

And then downgrade 3.9.0 back to 3.8.2 because the sweet success was
apparently an unreproducible mirage, and every subsequent attempt
fails just as described here:

https://www.mail-archive.com/dev@maven.apache.org/msg121747.html

... and observe sweet success again with 3.8.2, and hope it lasts.