Correct DOCUMENT v. CONTENT whitespace handling
There are more kinds of whitespace in XML than were dreamt of
in my (former) philosophy.

My uses of isElementContentWhitespace in domUnwrap, and of (SPACE = evt)
in useWrappingElement, were ill-informed. Both are specified to pertain
to a category of whitespace that can never exist where I was testing for
it.  "Element content whitespace" is a term of art in the spec, and
refers only to direct text-node children of elements that are explicitly
declared to have "element content", which can only happen in a DTD, and
there is no way to reach the code where I was using those tests if a DTD
is found in the input. Therefore nothing being checked for whitespace in
domUnwrap was matching (even when it was whitespace), and domUnwrap was
producing DocumentFragment in some cases when Document was correct.

"Element content whitespace" is a different category from whitespace
found at the top level, outside of any element. The XML specs declare
top-level whitespace (I could not make this up) to be "markup" and
not "character data" at all! While a parser that detects element
content whitespace (in an element that can have it, as declared in
the DTD) is supposed to pass it on to the application, suitably
labeled, the whitespace that is "markup" seems not to get reported
to the application at all.

In this commit, completely correct the handling of "markup"
whitespace in domUnwrap, and introduce a WhitespaceAccumulator
class to use for an efficiently-streamable approximation to
correct behavior for SAX and StAX. (Nothing is implemented in
the SAX- or StAX-related code yet; this just adds the needed
WhitespaceAccumulator class.)

The "efficiently-streamable approximation" will always have
correct behavior for input that turns out to be DOCUMENT, but
may lose some "markup" whitespace in input that turns out to be
CONTENT, preceding the point where it first becomes certain
that the input has CONTENT form.

That may be good enough to cover a lot of uses. Consumers most
likely to be thrown by "markup" whitespace may be those strictly
expecting DOCUMENT, and what they see will be strictly correct.

For the approximately-correct CONTENT case, it would be hard to
do better, as long as PostgreSQL does not save information on
the form of the value, and java.xml.SQLXML doesn't expose a way
for the application to weigh in. One could be added with
Adjusting.XML, perhaps. The determination of form can require
pre-parsing all the way to the end of the document just to decide
how to really parse it; that just seems excessive.