Apply the same whitespace fix in StAXUnwrapFilter
After working the same behavior into the SAX filter, just a simple
matter of turning brain inside out and hey presto! it's done for StAX.

There is one knuckle-biting moment in the StAX spec, here:

  int next()
  ...
  If the property javax.xml.stream.isCoalescing is set to true element
  content must be coalesced and only one CHARACTERS event must be
  returned for contiguous element content or CDATA Sections.

That would be more painful to implement. But a helpful language lawyer
points out that the requirement above applies explicitly to "element
content", and the only CHARACTERS events being given different treatment
here are the ones properly classified as "markup".