Fix a stored non-global-JNI-reference bug
Also arrange for Invocation's "boot context" to also push
and pop a JNI local reference frame, to unburden initPLJavaClasses
of tracking and deleting local references retail (as it is
not called *from* Java, that isn't guaranteed to be handled
automatically). Should also silence a couple local ref
"exceeds capacity" warnings, albeit not the thundering hordes
of them from the depths of JMX.

I was even expecting that it might flush out some more
bugs in reference management inside the boot context, but there
are no more complaints. It may have been just this one.