Dodge an excessively-strict -Xcheck:jni check
The various numeric sub'classes' of Type.c retain methodIDs
to the primitive-returning methods of Java's corresponding
boxed types. For instance, Double.c keeps a method reference
to doubleValue().

Those methods are really inherited or overridden methods
of Number, and PL/Java has been relying on that, as can be seen
in SPIConnection.basicNumericCoercion() being a no-op as long as
the value is assignable to Number, even if it isn't strictly
of the expected class. The assumption is that in Double.c, for
example, it will work to call doubleValue() even on an object
that isn't a Double, as long as it's a Number.

That works fine with -Xcheck:jni off, but when on, it's reported
as a FATAL ERROR if the methodID (obtained from Double.class)
is applied to anything that isn't assignable to Double.

The systematic way to fix that would be to rototill all the C
numeric Type 'classes' and have them get their fooValue()
methodIDs from Number.class rather than Foo.class.

But because this only seems to be a case of -Xcheck:jni being
more strict than JVM rules require, for now I will take a lazy
way out, and tweak the one case among all existing tests that
happens to trigger the error (the SELECT 1.5 with no explicit
cast has type NUMERIC by default, which is fetched as a Java
BigDecimal where the result tupledesc would call for a FLOAT,
i.e. a Java double).