Eat the local-ref capacity warnings from JMX
Backpatch of 4e001ae.

A lot of -Xcheck:jni output consists of local reference capacity
exceeded warnings coming from way down in the Java Management
Extensions where nothing in PL/Java's power to do would fix them.

The sheer volume of those makes it difficult to discern any real
PL/Java issues being reported among the JNI checks.

Because the -Xcheck:jni output goes through the vfprintf hook,
a simple state machine can be added there to recognize the
particular sequences of messages originating in the culprit JMX
class, and swallow them. Other messages will be allowed through.

This is quite hacky and sure to break in non-en_US locales (if
the -Xcheck:jni messages are localized; then it would simply fail
to recognize and eat them, and the noise would be back), but
should save a lot of fuss going through CI log output, as long
as the CI runs are done in en_US.

I was expecting to see a handful of actual PL/Java issues reported
once the flood of these was out of the way, but it seems the
earlier commits have taken care of those for now. I actually get
a quiet run (which unnerved me enough to step through the state
machine in gdb and make sure it wasn't eating anything it shouldn't,
and it wasn't).

There is no 'flush' operation ever applied to this state machine;
there could, in the absence of any following message or stack trace,
be one unreported "WARNING: JNI local refs: %u, exceeds capacity: %u"
line that doesn't get seen. That seems acceptable; it's just a
warning, and is usually followed by a stack trace, so the chance of
it not being flushed is slim. Sending any message at all through the
vfprintf hook that doesn't match that pattern will cause any saved
one to be flushed.