Make JNI local ref capacity warning more helpful
Backpatch of 6f63568.

It helps to know some things about tracking down these warnings.
The stacktrace given by the JVM is not terribly helpful, as it
goes only as deep as the native method called from Java, and not
the deeper calls it has made, so it is not uncommon to look at the
method suggested by the stacktrace and find nothing that could
give rise to the warning.

So display the capacity warning (if displayed at all, that is, not
the one from JMX) as an INFO ereport with just the first line of
its JVM-provided stacktrace as errdetail, and with an errhint
explaining where to set a gdb breakpoint and use that stacktrace
to find the the real offender.