Fix XMLTABLE for XML-typed columns and parameters
XML-typed output columns didn't work because of a simple
omitted test for the type.

XML-typed input parameters didn't work because of the (extension
to the spec) choice to make them available to column expressions
as well as to the row expression, while the JDBC SQLXML type is
only allowed to be read one time. It doesn't make sense to do the
work of tree building more than once anyway, so simply cache the
built tree rather than the SQLXML object.

Addresses #280.