Add the JAXP security-related properties
The properties, described in [1], are distinct from
the already supported 'features', which are simply on/off;
properties take values of specified types.

Only the ones applicable to XML parsers are added here; the
SQLXML API doesn't wrap or conceal instantiation of schemas
or transformers, so existing Java API is available for
configuring those.

[1] https://docs.oracle.com/en/java/javase/13/security/java-api-xml-processing-jaxp-security-guide.html