Simplify
The defaults() methods are the same everywhere and can be hoisted
into AdjustingJAXPParser.

Several adjustments have to be done differently in StAX but at least
are common to SAX and DOM, so can be hoisted into a subclass that
SAX and DOM extend.

Once Java 8 is baseline, many of these can simply be default methods
in the API interfaces.