Handle set(SAXSource) when no XMLReader supplied
Introduction of the Adjusting.XML.Parsing API left one non-obvious
case where a parser could be created internally without being exposed
for adjustments through the new API.

When AdjustingSourceResult.set(SAXSource) is passed a SAXSource that
wraps an InputSource but no XMLReader of its own, the expectation is
that an XMLReader will be created and used to parse the InputSource.
This case is effectively a stream source but passed via SAX API.

So create the XMLReader, but make sure it is exposed as adjustable.
No special effort is made to 'correct' the XML Decl, wrap for
possible CONTENT form, etc.; if some foreign driver is handing us
a no-XMLReader SAXSource and expecting us to parse it, it had better
be acceptable to a vanilla JAXP parser.