Fix function/prototype mismatch
There were two issues here. That the prototype and function had
the parameters in a different order was not a compile error
(because the types were the same), but certainly misleading
and a risk for future maintenance.

That the types were int in the prototype and jint in the function
is a compile error, but has never been flagged by any compiler
until building with Mingw-w64 (where presumably there is more of
a distinction between jint and int).

Addresses #282.

Was introduced in 1.5.3 so should be backpatched to 1.5.