Refactor for easier alternatives to ScriptEngines
With the removal of Nashorn in Java 15, leaving no scripting language
that can be assumed present in the Java runtime, a script in the
manifest may no longer be the simplest way to customize the resolution
of path names when extracting. This class has been refactored now to
expose two methods, prepareResolver and resolve, easily overridden in a
subclass.  The value of the _JarX_PathResolver main attribute is passed
to prepareResolver as a string (so it can be parsed in any way useful to
the subclass, not necessarily as described in JarX javadoc, or ignored),
and resolve is passed the stored path and platform path, and returns the
platform path unchanged or a replacement. A self-extracting jar with
resolution can be made without depending on any script engine, by
placing /two/ classes in the jar, JarX and the subclass, and naming the
subclass as the jar's Main-Class. It needs a main method that simply
instantiates the class and calls extract().