Add graaljs dependencies if built on Java 15+
Having a usable scripting language batteries-included in the
Java runtime has been very helpful for many years, but Nashorn
no longer comes in the box starting with Java 15. There is a
functional JavaScript engine available (from the GraalVM project,
but the engine itself will run on a stock JVM), and even supports
much more recent JavaScript language features (that we're not using).

So in a profile detecting Java 15 or later, add dependencies for the
maven-antrun-plugin on org.graalvm.js:js (the language implementation
proper, all sixteen magnificent megabytes of it), and on
org.graalvm.js:js-scriptengine (the 57 kB shim to make it appear as
an available javax.script engine, as ant expects). Also set the
system property polyglot.js.nashorn-compat so the Nashorn-isms in
our existing JavaScript code continue to work.