Setup PL/Java PGXS plugin
This commit intriduces the PL/Java PGXS plugin. The ultimate
aim is to replace nar maven plugin in PL/Java's build system
to solve the fair share of problems it brings with it. All big
things begin small. This commit makes it possible for the PGXS
plugin to launch pg_config, retrieve the required properties
from it and set those in the maven project. These properties
are required by nar maven plugin during its build process.
Before this commit, the task was achieved using maven
properties and maven antrun plugins. The commit also removes
these from the relevant parts where they are no longer needed.

The pljava build process already uses bits of
javascript. The aim is to extract the bland
parts of JavaScript in the Java Code and expose
a nice set of functions that make configuring
the build process from the pom.xml using small
javascript snippets easy. This commit is the
first step in that direction.

Instead of the setting the Properties using
PropertiesMojo and PackagingMojo, now a
setProjectProperty and getPgConfigProperty
method is exposed by the ScriptingMojo. These
methods are called by in a script block inside
th configuration block of the plugin in pom.xml.