Start on PostgresNode-like behavior
Add some methods patterned on PostgreSQL's PostgresNode.pm ...
just enough to allow using jshell with this jar to initialize /
start / stop / etc. a PG instance for testing.

Gotta run it like this:

jshell \
--class-path=pljava-packaging/target/pljava-pg....jar \
--execution=local \
-J--class-path=pljava-packaging/target/pljava-pg....jar:\
-J-Dpgconfig=path/to/pgversion/bin/pg_config

The first --class-path is for the jshell REPL, so it will let you
talk about the classes in the jar. Then --execution=local tells jshell
to run your snippets in the same JVM it is running in itself. (In its
default "jdi" execution style, a separate JVM is started to run your
snippets, and standard output/error *from that JVM* are brought back
to jshell's own output where you can see them, but output/error *from
subprocesses it spawns* are lost.)

The remaining -J options are needed because of the local execution.
The earlier --class-path option only controlled what jshell
believes is the class path, but -J--class-path is also needed so the
underlying JVM has that path too.

Finally, the -J-Dpgconfig= is where you point to the pg_config binary
for the PostgreSQL version you want to use.

A -J-Djava.io.tmpdir= can be used if the default temporary directory
isn't what's wanted.