Add ability to connect to the node
For convenience, a new Maven profile pgjdbc-ng can be specified when
building, and will declare the latest version of the pgjdbc-ng driver
as a dependency. It is not really needed for the build, but declaring
it a dependency will conveniently get it downloaded into the local
Maven repository, where it can be added to jshell's class path and
used to do some testing.

Serving suggestion:

jshell \
--class-path=pljava-packaging/target/pljava-pg....jar \
--execution=local \
-J--add-modules=java.sql,java.sql.rowset \
-J--class-path=pljava-packaging/target/pljava-pg....jar:\
$HOME/.m2/repository/com/impossibl/pgjdbc-ng/pgjdbc-ng-all/.../pgjdbc-ng-all-....jar \
-J-Dcom.impossibl.shadow.io.netty.noUnsafe=true \
-J-Dpgconfig=path/to/pgversion/bin/pg_config

The -J--add-modules is needed because even though jshell treats java.sql
and java.sql.rowset as available by default, the local JVM it is running on
wouldn't know that without being told.

The class path given to the JVM also needs to mention the pgjdbc-ng-all
jar if it will be used to make connections to a node. Don't waste time
with the original pgjdbc, it completely throws away notices/warnings
so isn't very usable for testing purposes. (As for pgjdbc-ng, it will
collect whatever warnings, notices, infos it receives and chain them all
together as SQLWarnings. To tell them apart, use getSQLState(); it should
be 00000 for non-warnings. Which of those it does or doesn't receive,
of course, is controlled by client_min_messages.)

The path given to jshell itself does not need that jar, because
pgjdbc-ng can be just a provider of the java.sql.Driver service
and doesn't have to be visible.

The noUnsafe setting silences a complaint from the netty library
about Java (correctly!) denying it access to private internals.

A number of convenience methods are added to use an existing JDBC
connection and: change PostgreSQL config settings, CREATE EXTENSION
pljava, install and remove jars, set the jar path, and so on.

Aiming to unify, as nearly as practical, the interface to
resultset-returning, count-returning, and nothing-returning
SQL snippets, without forgetting to check for warnings/notices,
pretty much all of the convenience methods return a Stream<Object>.
The objects coming back on the stream can be ResultSet instances,
Long instances (for commands that return update counts), SQLException
or SQLWarning instances. (An SQLException or SQLWarning can be the
head of a list of such, and has its own iterator/forEach methods that
should be used to traverse the whole list.)

A roughly thrown-together print() method is provided to take such a
Stream<Object> and print whatever exceptions/warnings/update counts/
ResultSet instances are retrieved. A ResultSet is dumped in the XML
format produced by WebRowSet. It is unpleasingly verbose, and does
not support all data types well, but at least is a way to dump a
ResultSet in fairly complete form using few lines of code.