Add some combined, convenience methods
Add methods to install the examples jar and place it on the
class path, to install Saxon from an expected location, to
install Saxon and place it on the class path, or a four-fer
that does all of the above in the right order.

These all return concatenated result streams from their
constituent operations. That probably would not work nearly
as well if some of those operations had streams of data to return.
But as they all return void, pgjdbc-ng does not seem to have any
problem with the consecutive Statements remaining open while their
streams are consumed, and it's quite convenient.