Don't let failed Statements dangle
It's a bit subtle to manage the Statements created within
convenience methods: they can't be closed immediately, as if
in a try-with-resources, because when the method returns a
stream of results, the Statement has to remain open as long
as the stream does. But if an exception is thrown on execution
of the statement, before the result stream is created, the
statement does have to be closed.

For consistency, now any SQLException thrown during that
initial execution of the statement is returned in the result
stream, as would any later exception thrown while retrieving
the results. Any Throwable outside the SQLException hierarchy,
on the other hand, is simply rethrown.