Simplify, favoring short method names
Once there was an overload qp(Stream<Object>) as a convenient
short alias for print(Stream<Object>), it was a small step to see
that the method should just be qp; there's no need for an alias
with a longer name. In fact, just overload qp to print all the
things you might want to print, and q for the ways you might want
to query. Two static imports, q and qp, will take you a long way.

Axe the createExtensionPLJava method. It was hardly shorter, and
no clearer, than qp(c, "create extension pljava").

Factor a classify() method out of qp(Throwable), to simplify code
that wants to classify them the same way for analysis rather than
printing.