Add a qp overload to examine ParameterMetaData
Continuing in the spirit of least work to get something
that just clears the bar of 'usable', this simply fakes up
a ResultSetMetaData with the values found in the supplied
ParameterMetaData (leaving unset all the details, like column
names, that ResultSetMetaData has but not ParameterMetaData),
then uses writeXml again as if dumping a ResultSet.

For a quick answer to the question "hey, what does PostgreSQL
think the parameters to this statement should be?", it's
good enough.