Make Checked more nearly orthogonal
Complete the set of ToFooFunction interfaces, for Foo in
byte, short, char, float, plus Predicate, so that the reference
and all primitive types are covered; complete the in(...) methods
to cover the full set of those functions.

Let the new functional interfaces that do not have Java API
checked-exception-less counterparts, and therefore do not strictly
need to have ederWrap methods, inherit a no-op default one anyway.
That eliminates arbitrary limits on which ones can participate in
the use(...).in(...) idiom.