Try an external PowerShell script
As suggested by Feodor Fitsner in [1], maybe the behavior of
a separate script will be better than when the script is in
the .yml file.

The staff seems to be strangely cagey about exactly what's
happening that makes this stuff not work. There are other threads on
the same topic (e.g. [2]). "you can try playing with StdErr
redirection ... but ... not always work" ?!  Seriously?

[1] https://help.appveyor.com/discussions/questions/48697#comment_48398341
[2] https://help.appveyor.com/discussions/problems/10014#comment_44099248