Will it run that far, put back in appveyor.yml?
On the last run it got as far as postgres's "I refuse to run
as an adminstrative user" check. It will take some more doing
to get past that. There seem to be relevant AppVeyor threads
in [1] and [2].

But anyway, meanwhile, can it be put back into appveyor.yml,
doing away with the separate .ps1 file, and still succeed as far?

[1] https://help.appveyor.com/discussions/questions/1888-running-tests-with-reduced-privileges
[2] https://help.appveyor.com/discussions/questions/3530-is-it-possible-to-somehow-specify-the-usernamehostname-with-which-ci-tests-are-run