Try a 2-parent classloader for ScriptEngineManager
My understanding (since 6cd539b) has been that Java 9 moved Nashorn
from the bootstrap classloader to the application classloader [1], and
that Maven creates classloaders that do not delegate to the application
classloader (a/k/a the system classloader, as various ClassLoader
methods still call it).

So, what seemed an easy solution would be to create a new ClassLoader
as a child of the Maven-supplied loader, but that would also behave as
if it had the application class loader as a second parent.

It still doesn't find Nashorn though. Back to the drawing board.

[1] https://issues.apache.org/jira/browse/MANTRUN-200?focusedCommentId=16138712&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-16138712